refactor(images): extract create() into ImageGenerationService, dedup provider scaffolding (#778)
images-operations.controller.ts was 1,494 lines dominated by a ~1,090-line
create() with 27 injected deps. Extract the full image-generation workflow into
a new ImageGenerationService; the controller's create() is now a thin handler
that keeps all decorators/guards/interceptors and delegates. The controller
constructor drops to 8 deps (split-endpoint only + the service). Mirrors the
videos #835 VideoGenerationService precedent.

Decomposition:
- Provider selection is a dispatch map (classifyProvider -> dispatchGeneration
  switch), not a flat if-chain. Adding a provider touches only a handler + the
  switch.
- The 5 copy-pasted failedGenerationService.handleFailedImageGeneration calls
  collapse to one handleProviderFailure().
- The 3 duplicated PollingTimeoutError recovery blocks collapse to one
  throwGatewayTimeoutIfPending().
- The poll/serialize tails collapse into finishGeneration() +
  resolveCompletedIngredient(), keyed by a CompletionPlan kind that preserves
  each provider's exact wait semantics (GenfeedAi inline / Kling+Leonardo
  poll-single / Replicate poll-multiple / Fal background-only / SDXL none).
- The placeholder activity+websocket triplication collapses to one
  createImagePlaceholderActivity().
- The shared mutable `url` label is gone; each provider handler passes its own.

Behavior-preserving: routes, decorators, guards, status codes
(BAD_REQUEST/FORBIDDEN/PAYMENT_REQUIRED/GATEWAY_TIMEOUT), JSON:API response
shape, and background-vs-wait behavior unchanged. Every method in both files,
except the pre-existing out-of-scope splitContactSheet, is now <=150 lines. Spec
rewired to construct the real ImageGenerationService from the same mocks; all 38
assertions unchanged and passing.

Closes #778
