+++ apps/server/api/src/collections/images/controllers/operations/images-operations.controller.ts
@@ -1,28 +1,11 @@
 import type { AuthenticatedUser as User } from '@api/auth/interfaces/authenticated-user.interface';
 import { ActivityEntity } from '@api/collections/activities/entities/activity.entity';
 import { ActivitiesService } from '@api/collections/activities/services/activities.service';
-import { AssetsService } from '@api/collections/assets/services/assets.service';
-import { BrandsService } from '@api/collections/brands/services/brands.service';
-import { buildPromptBrandingFromBrand } from '@api/collections/brands/utils/brand-context.util';
-import { CreditsUtilsService } from '@api/collections/credits/services/credits.utils.service';
 import { CreateImageDto } from '@api/collections/images/dto/create-image.dto';
 import { SplitImageDto } from '@api/collections/images/dto/split-image.dto';
+import { ImageGenerationService } from '@api/collections/images/services/image-generation.service';
 import { ImagesService } from '@api/collections/images/services/images.service';
 import type { IngredientRefDocument } from '@api/collections/ingredients/schemas/ingredient.schema';
-import { IngredientsService } from '@api/collections/ingredients/services/ingredients.service';
-import { MetadataEntity } from '@api/collections/metadata/entities/metadata.entity';
-import { MetadataService } from '@api/collections/metadata/services/metadata.service';
-import { ModelRegistrationService } from '@api/collections/models/services/model-registration.service';
-import { ModelsService } from '@api/collections/models/services/models.service';
-import {
-  baseModelKey,
-  isFalDestination,
-  isGenfeedAiDestination,
-  isReplicateDestination,
-} from '@api/collections/models/utils/model-key.util';
-import { OrganizationSettingsService } from '@api/collections/organization-settings/services/organization-settings.service';
-import { PromptEntity } from '@api/collections/prompts/entities/prompt.entity';
-import { PromptsService } from '@api/collections/prompts/services/prompts.service';
 import { CreateTagDto } from '@api/collections/tags/dto/create-tag.dto';
 import { TagsService } from '@api/collections/tags/services/tags.service';
 import type { RequestWithContext as Request } from '@api/common/middleware/request-context.middleware';
@@ -43,30 +26,12 @@ import { RolesGuard } from '@api/helpers/guards/roles/roles.guard';
 import { SubscriptionGuard } from '@api/helpers/guards/subscription/subscription.guard';
 import { CreditsInterceptor } from '@api/helpers/interceptors/credits/credits.interceptor';
 import { getPublicMetadata } from '@api/helpers/utils/auth/auth.util';
-import { resolveGenerationDefaultModel } from '@api/helpers/utils/generation-defaults/generation-defaults.util';
-import { buildReferenceImageUrls } from '@api/helpers/utils/reference/reference.util';
-import { serializeSingle } from '@api/helpers/utils/response/response.util';
-import { WebSocketPaths } from '@api/helpers/utils/websocket/websocket.util';
 import { isEntityId } from '@api/helpers/validation/entity-id.validator';
 import { FilesClientService } from '@api/services/files-microservice/client/files-client.service';
-import { ComfyUIService } from '@api/services/integrations/comfyui/comfyui.service';
-import { FalService } from '@api/services/integrations/fal/fal.service';
-import { KlingAIService } from '@api/services/integrations/klingai/klingai.service';
-import { LeonardoAIService } from '@api/services/integrations/leonardoai/leonardoai.service';
-import { ReplicateService } from '@api/services/integrations/replicate/replicate.service';
-import { NotificationsService } from '@api/services/notifications/notifications.service';
-import { NotificationsPublisherService } from '@api/services/notifications/publisher/notifications-publisher.service';
-import { PromptBuilderService } from '@api/services/prompt-builder/prompt-builder.service';
-import { RouterService } from '@api/services/router/router.service';
-import { WebhookClientService } from '@api/services/webhook-client/webhook-client.service';
 import { RateLimit } from '@api/shared/decorators/rate-limit/rate-limit.decorator';
-import { FailedGenerationService } from '@api/shared/services/failed-generation/failed-generation.service';
-import { PollingService } from '@api/shared/services/polling/polling.service';
 import { SharedService } from '@api/shared/services/shared/shared.service';
 import { PopulatePatterns } from '@api/shared/utils/populate/populate.util';
-import { MODEL_KEYS, MODEL_OUTPUT_CAPABILITIES } from '@genfeedai/constants';
 import {
-  ActivityEntityModel,
   ActivityKey,
   ActivitySource,
   FileInputType,
@@ -75,17 +40,11 @@ import {
   MemberRole,
   MetadataExtension,
   ModelCategory,
-  PricingType,
-  PromptCategory,
-  PromptStatus,
   TagCategory,
   TagKey,
 } from '@genfeedai/enums';
 import type { JsonApiSingleResponse } from '@genfeedai/interfaces';
-import { IngredientSerializer } from '@genfeedai/serializers';
 import { LoggerService } from '@libs/logger/logger.service';
-import { getErrorMessage } from '@libs/utils/error/get-error-message.util';
-import { getUserRoomName } from '@libs/websockets/room-name.util';
 import {
   Body,
   Controller,
@@ -105,37 +64,15 @@ import sharp from 'sharp';
 @Controller('images')
 @UseGuards(RolesGuard)
 export class ImagesOperationsController {
-  private readonly constructorName: string = String(this.constructor.name);
-
   constructor(
     private readonly configService: ConfigService,
     private readonly activitiesService: ActivitiesService,
-    private readonly assetsService: AssetsService,
-    private readonly brandsService: BrandsService,
-    private readonly comfyUIService: ComfyUIService,
-    private readonly creditsUtilsService: CreditsUtilsService,
-    private readonly failedGenerationService: FailedGenerationService,
     private readonly filesClientService: FilesClientService,
-    private readonly falService: FalService,
-    private readonly pollingService: PollingService,
     private readonly imagesService: ImagesService,
-    private readonly ingredientsService: IngredientsService,
-    private readonly organizationSettingsService: OrganizationSettingsService,
-    private readonly klingAIService: KlingAIService,
-    private readonly leonardoaiService: LeonardoAIService,
     private readonly loggerService: LoggerService,
-    private readonly metadataService: MetadataService,
-    private readonly modelRegistrationService: ModelRegistrationService,
-    private readonly modelsService: ModelsService,
-    readonly _notificationsService: NotificationsService,
-    private readonly promptBuilderService: PromptBuilderService,
-    private readonly promptsService: PromptsService,
-    private readonly replicateService: ReplicateService,
-    private readonly routerService: RouterService,
     private readonly sharedService: SharedService,
     private readonly tagsService: TagsService,
-    readonly _webhookService: WebhookClientService,
-    private readonly websocketService: NotificationsPublisherService,
+    private readonly imageGenerationService: ImageGenerationService,
   ) {}
 
   @Post()
@@ -160,1091 +97,11 @@ export class ImagesOperationsController {
     @Body() createImageDto: CreateImageDto,
     @CurrentUser() user: User,
   ): Promise<JsonApiSingleResponse> {
-    let url = `${this.constructorName} create`;
-
-    this.loggerService.log(url, { ...createImageDto });
-    const publicMetadata = getPublicMetadata(user);
-
-    if (!createImageDto.prompt && !createImageDto.text) {
-      throw new HttpException(
-        {
-          detail: 'Prompt is required',
-          title: 'Prompt validation failed',
-        },
-        HttpStatus.BAD_REQUEST,
-      );
-    }
-
-    const promptOriginal = createImageDto.text || createImageDto.prompt;
-    const promptOriginalText =
-      typeof promptOriginal === 'string'
-        ? promptOriginal
-        : String(promptOriginal ?? '');
-
-    const brandId = createImageDto.brand || publicMetadata.brand;
-    const brand = await this.brandsService.findOne({
-      _id: brandId,
-      isDeleted: false,
-      organization: publicMetadata.organization,
-    });
-
-    if (!brand) {
-      throw new HttpException(
-        {
-          detail: 'You do not have access to this brand',
-          title: 'Brand not found',
-        },
-        HttpStatus.FORBIDDEN,
-      );
-    }
-
-    const organizationSettings = await this.organizationSettingsService.findOne(
-      {
-        isDeleted: false,
-        organization: publicMetadata.organization,
-      },
+    return this.imageGenerationService.generateImage(
+      user,
+      createImageDto,
+      request,
     );
-
-    // Model selection: auto-select > user-provided > brand default > system default
-    let model: string;
-    let routerReason: string | undefined;
-
-    if (createImageDto.autoSelectModel) {
-      // Auto model routing - let RouterService pick the best model
-      const promptString = promptOriginalText;
-      const recommendation = await this.routerService.selectModel({
-        category: ModelCategory.IMAGE,
-        dimensions: {
-          height: createImageDto.height,
-          width: createImageDto.width,
-        },
-        outputs: createImageDto.outputs,
-        prioritize: createImageDto.prioritize || 'balanced',
-        prompt: promptString,
-      });
-      model = recommendation.selectedModel as string;
-      routerReason = recommendation.reason;
-
-      this.loggerService.log('Auto model routing selected', {
-        promptPreview: promptString.substring(0, 100),
-        reason: routerReason,
-        selectedModel: model,
-        service: this.constructorName,
-      });
-    } else {
-      // Manual selection: user-provided > brand default > system default
-      const modelKeys = Object.values(MODEL_KEYS) as string[];
-      const userModel =
-        createImageDto.model &&
-        modelKeys.includes(createImageDto.model as string)
-          ? (createImageDto.model as string)
-          : undefined;
-      const brandDefaultModel =
-        brand.defaultImageModel &&
-        modelKeys.includes(brand.defaultImageModel as string)
-          ? (brand.defaultImageModel as string)
-          : undefined;
-      const organizationDefaultModel =
-        organizationSettings?.defaultImageModel &&
-        modelKeys.includes(organizationSettings.defaultImageModel as string)
-          ? (organizationSettings.defaultImageModel as string)
-          : undefined;
-      const systemDefaultModel = (await this.routerService.getDefaultModel(
-        ModelCategory.IMAGE,
-      )) as string;
-      model = resolveGenerationDefaultModel<string>({
-        brandDefault: brandDefaultModel,
-        explicit: userModel,
-        organizationDefault: organizationDefaultModel,
-        systemDefault: systemDefaultModel,
-      });
-    }
-
-    // Validate resolved model against org (catches default-resolution bypassing ModelsGuard)
-    if (request.context?.organizationId) {
-      const authenticatedOrgId = request.context.organizationId;
-      await this.modelRegistrationService.validateModelForOrg(
-        model,
-        authenticatedOrgId,
-      );
-    }
-
-    // CreditsGuard deferred credit check, do it now with resolved model.
-    const reqWithCredits = request as unknown as {
-      creditsConfig?: {
-        deferred?: boolean;
-        amount?: number;
-        modelKey?: string;
-      };
-    };
-    if (reqWithCredits.creditsConfig?.deferred) {
-      const resolvedModelDoc = await this.modelsService.findOne({
-        isDeleted: false,
-        key: baseModelKey(model),
-      });
-
-      const imgWidth = createImageDto.width || 1920;
-      const imgHeight = createImageDto.height || 1080;
-      let requiredCredits: number;
-
-      if (resolvedModelDoc) {
-        requiredCredits = this.calculateDynamicImageCost(
-          resolvedModelDoc,
-          imgWidth,
-          imgHeight,
-        );
-      } else {
-        requiredCredits = 5; // Fallback default cost
-      }
-
-      const hasCredits =
-        await this.creditsUtilsService.checkOrganizationCreditsAvailable(
-          publicMetadata.organization,
-          requiredCredits,
-        );
-      if (!hasCredits) {
-        const balance =
-          await this.creditsUtilsService.getOrganizationCreditsBalance(
-            publicMetadata.organization,
-          );
-        throw new HttpException(
-          {
-            detail: `Insufficient credits: ${requiredCredits} required, ${balance} available`,
-            title: 'Insufficient credits',
-          },
-          HttpStatus.PAYMENT_REQUIRED,
-        );
-      }
-      reqWithCredits.creditsConfig = {
-        ...reqWithCredits.creditsConfig,
-        amount: requiredCredits,
-        deferred: false,
-        modelKey: model,
-      };
-    }
-
-    const replicateModels: string[] = [
-      MODEL_KEYS.REPLICATE_GOOGLE_IMAGEN_3,
-      MODEL_KEYS.REPLICATE_GOOGLE_IMAGEN_4,
-      MODEL_KEYS.REPLICATE_GOOGLE_IMAGEN_4_FAST,
-      MODEL_KEYS.REPLICATE_GOOGLE_IMAGEN_4_ULTRA,
-    ];
-
-    const isLeonardoAI = model === MODEL_KEYS.LEONARDOAI;
-    const isKlingAI = model === MODEL_KEYS.KLINGAI_V2;
-    const isSDXL = model === MODEL_KEYS.SDXL;
-    const isGenfeedAi = isGenfeedAiDestination(model);
-    const isFal = isFalDestination(model);
-    const isKnownReplicateModel = replicateModels.includes(model);
-    const isDynamicReplicateDestination = isReplicateDestination(model);
-    const isReplicate =
-      !isFal &&
-      !isGenfeedAi &&
-      (isKnownReplicateModel || isDynamicReplicateDestination);
-
-    const brandPromptBranding = buildPromptBrandingFromBrand(brand);
-    const promptBuilderBrand = {
-      description: brand.description ?? undefined,
-      label: brand.label ?? 'Brand',
-      primaryColor: brand.primaryColor ?? undefined,
-      secondaryColor: brand.secondaryColor ?? undefined,
-      text: brand.text ?? undefined,
-    };
-
-    const width = createImageDto.width || 1920;
-    const height = createImageDto.height || 1080;
-    const style = createImageDto.style;
-    const outputs = Number(createImageDto.outputs) || 1;
-
-    this.loggerService.debug('Image generation request received', {
-      model,
-      outputs,
-      rawOutputs: createImageDto.outputs,
-    });
-
-    // Build reference URLs from references array
-    const referenceIds: string[] = Array.isArray(createImageDto.references)
-      ? createImageDto.references.map((id) => id.toString())
-      : [];
-
-    const referenceImageUrls: string[] = await buildReferenceImageUrls({
-      assetsService: this.assetsService,
-      configService: this.configService,
-      ingredientsService: this.ingredientsService,
-      loggerService: this.loggerService,
-      referenceIds,
-    });
-
-    const referenceImageUrl: string | null = referenceImageUrls[0] || null;
-
-    if (
-      !isLeonardoAI &&
-      !isKlingAI &&
-      !isSDXL &&
-      !isGenfeedAi &&
-      !isFal &&
-      !isReplicate
-    ) {
-      throw new HttpException(
-        {
-          detail: 'Invalid model for image generation',
-          title: 'Validation failed',
-        },
-        HttpStatus.BAD_REQUEST,
-      );
-    }
-
-    const promptData = await this.promptsService.create(
-      new PromptEntity({
-        brand: isEntityId(createImageDto.brand)
-          ? createImageDto.brand
-          : publicMetadata.brand,
-        category: PromptCategory.MODELS_PROMPT_IMAGE,
-        model: createImageDto.model as string,
-        organization: publicMetadata.organization,
-        original: promptOriginalText,
-        status: PromptStatus.PROCESSING,
-        user: publicMetadata.user,
-      }),
-    );
-
-    // Build prompt early to get template tracking info
-    const {
-      templateUsed: imageTemplateUsed,
-      templateVersion: imageTemplateVersion,
-    } = await this.promptBuilderService.buildPrompt(
-      model,
-      {
-        blacklist: createImageDto.blacklist,
-        brand: promptBuilderBrand,
-        branding: brandPromptBranding,
-        brandingMode: createImageDto.brandingMode,
-        camera: createImageDto.camera,
-        fontFamily: createImageDto.fontFamily,
-        height,
-        isBrandingEnabled: createImageDto.isBrandingEnabled,
-        lens: createImageDto.lens,
-        lighting: createImageDto.lighting,
-        // Use model's category from DB (set by ModelsGuard), fallback to IMAGE
-        modelCategory: ModelCategory.IMAGE,
-        mood: createImageDto.mood,
-        outputs: createImageDto.outputs,
-        prompt: promptData.original,
-        promptTemplate: createImageDto.promptTemplate,
-        references: referenceImageUrls,
-        scene: createImageDto.scene,
-        seed: createImageDto.seed,
-        style: style || createImageDto.style || 'realistic',
-        useTemplate: createImageDto.useTemplate,
-        width,
-      },
-      publicMetadata.organization,
-    );
-
-    const { metadataData, ingredientData } =
-      await this.sharedService.saveDocuments(user, {
-        ...createImageDto,
-        brand: brand._id,
-        category: IngredientCategory.IMAGE,
-        extension: MetadataExtension.JPEG,
-        height,
-        model,
-        organization: publicMetadata.organization,
-        parent: isEntityId(createImageDto.parent)
-          ? createImageDto.parent
-          : undefined,
-        prompt: promptData._id,
-        // Template tracking
-        promptTemplate: imageTemplateUsed,
-        style,
-        templateVersion: imageTemplateVersion,
-        width,
-      });
-
-    await this.imagesService.patch(ingredientData._id, {
-      prompt: promptData._id,
-    });
-
-    // Create activity for image generation start (after ingredientData is available)
-    const activity = await this.activitiesService.create(
-      new ActivityEntity({
-        brand: brand._id,
-        entityId: ingredientData._id,
-        entityModel: ActivityEntityModel.INGREDIENT,
-        key: ActivityKey.IMAGE_PROCESSING,
-        organization: publicMetadata.organization,
-        source: ActivitySource.IMAGE_GENERATION,
-        user: publicMetadata.user,
-        value: JSON.stringify({
-          ingredientId: ingredientData._id.toString(),
-          model,
-          type: 'generation',
-        }),
-      }),
-    );
-
-    // Emit background-task-update WebSocket event for activities dropdown
-    await this.websocketService.publishBackgroundTaskUpdate({
-      activityId: activity._id.toString(),
-      label: 'Image Generation',
-      progress: 0,
-      room: getUserRoomName(user.id),
-      status: 'processing',
-      taskId: ingredientData._id.toString(),
-      userId: user.id,
-    });
-
-    const websocketUrl = WebSocketPaths.image(ingredientData._id);
-    const waitForCompletion = createImageDto.waitForCompletion === true;
-    const pendingIngredientIds: string[] = [ingredientData._id.toString()];
-
-    if (isGenfeedAi) {
-      url = 'ComfyUIService generateImage';
-
-      const generationPromise = (async () => {
-        const { imageBuffer } = await this.comfyUIService.generateImage(model, {
-          faceImage: referenceImageUrl || undefined,
-          height,
-          prompt: promptData.original,
-          seed: createImageDto.seed,
-          width,
-        });
-
-        const uploadMeta = await this.filesClientService.uploadToS3(
-          ingredientData._id.toString(),
-          'images',
-          {
-            contentType: 'image/png',
-            data: imageBuffer,
-            type: FileInputType.BUFFER,
-          },
-        );
-
-        await Promise.all([
-          this.metadataService.patch(
-            metadataData._id,
-            new MetadataEntity({
-              height: uploadMeta.height,
-              prompt: promptData._id,
-              size: uploadMeta.size,
-              width: uploadMeta.width,
-            }),
-          ),
-          this.imagesService.patch(ingredientData._id, {
-            cdnUrl:
-              typeof uploadMeta.publicUrl === 'string'
-                ? uploadMeta.publicUrl
-                : undefined,
-            prompt: promptData._id,
-            s3Key:
-              typeof uploadMeta.s3Key === 'string'
-                ? uploadMeta.s3Key
-                : undefined,
-            status: IngredientStatus.GENERATED,
-          }),
-          this.websocketService.publishVideoComplete(
-            websocketUrl,
-            {
-              id: ingredientData._id.toString(),
-              ingredientId: ingredientData._id.toString(),
-              status: 'completed',
-            },
-            user.id,
-            getUserRoomName(user.id),
-          ),
-        ]);
-
-        return ingredientData._id.toString();
-      })().catch(async (error: unknown) => {
-        this.loggerService.error(`${url} failed`, error);
-        const errorMessage = getErrorMessage(error);
-
-        await this.failedGenerationService.handleFailedImageGeneration(
-          this.imagesService,
-          ingredientData._id,
-          websocketUrl,
-          publicMetadata,
-          getUserRoomName(user.id),
-          errorMessage,
-        );
-        throw error;
-      });
-
-      if (waitForCompletion) {
-        await generationPromise;
-
-        const completedIngredient = await this.imagesService.findOne(
-          { _id: ingredientData._id },
-          [
-            PopulatePatterns.promptFull,
-            PopulatePatterns.metadataFull,
-            PopulatePatterns.brandMinimal,
-          ],
-        );
-
-        return serializeSingle(
-          request,
-          IngredientSerializer,
-          completedIngredient,
-        );
-      }
-
-      generationPromise.catch(() => {
-        // Error already handled in the catch block above
-      });
-    }
-
-    // Kling AI
-    if (isKlingAI) {
-      url = 'KlingAIService generateImage';
-
-      // Create async generation promise - don't await unless waitForCompletion
-      const generationPromise = this.klingAIService
-        .queueGenerateImage(promptData.original, {
-          ...createImageDto,
-          height,
-          model,
-          reference: referenceImageUrl || undefined,
-          style: style || 'realistic',
-          width,
-        })
-        .then(async (generationId) => {
-          if (!generationId) {
-            throw new Error('No generation ID returned from KlingAI');
-          }
-
-          await this.metadataService.patch(
-            metadataData._id,
-            new MetadataEntity({
-              externalId: generationId,
-              prompt: promptData._id,
-            }),
-          );
-
-          return generationId;
-        })
-        .catch(async (error: unknown) => {
-          this.loggerService.error(`${url} failed`, error);
-          const errorMessage = getErrorMessage(error);
-
-          await this.failedGenerationService.handleFailedImageGeneration(
-            this.imagesService,
-            ingredientData._id,
-            websocketUrl,
-            publicMetadata,
-            getUserRoomName(user.id),
-            errorMessage,
-          );
-          throw error;
-        });
-
-      if (waitForCompletion) {
-        // Wait for the generation to complete
-        try {
-          await generationPromise;
-
-          const completedIngredient =
-            await this.pollingService.waitForIngredientCompletion(
-              ingredientData._id.toString(),
-              180000, // 3 minutes timeout
-              2000, // 2 seconds poll interval
-              [
-                PopulatePatterns.promptFull,
-                PopulatePatterns.metadataFull,
-                PopulatePatterns.brandMinimal,
-              ],
-            );
-
-          return serializeSingle(
-            request,
-            IngredientSerializer,
-            completedIngredient,
-          );
-        } catch (error: unknown) {
-          if (error instanceof Error && error.name === 'PollingTimeoutError') {
-            const ingredient = await this.imagesService.findOne(
-              { _id: ingredientData._id },
-              [
-                PopulatePatterns.promptFull,
-                PopulatePatterns.metadataFull,
-                PopulatePatterns.brandMinimal,
-              ],
-            );
-
-            if (ingredient) {
-              throw new HttpException(
-                {
-                  detail: `Image generation did not complete within 3 minutes. Current status: ${ingredient.status}`,
-                  title: 'Generation timeout',
-                },
-                HttpStatus.GATEWAY_TIMEOUT,
-              );
-            }
-          }
-          throw error;
-        }
-      } else {
-        // Return immediately - generation runs in background
-        // Attach empty catch to prevent unhandled promise rejection
-        generationPromise.catch(() => {
-          // Error already handled in the catch block above
-        });
-      }
-    }
-
-    if (isFal) {
-      url = 'FalService generateImage';
-
-      const buildFalInput = (): Record<string, unknown> => ({
-        image_size: {
-          height,
-          width,
-        },
-        ...(referenceImageUrl ? { image_url: referenceImageUrl } : {}),
-        ...(createImageDto.seed !== undefined
-          ? { seed: createImageDto.seed }
-          : {}),
-        prompt: promptData.original,
-      });
-
-      const generationPromise = (async () => {
-        const falResult = await this.falService.generateImage(
-          model,
-          buildFalInput(),
-        );
-
-        await this.metadataService.patch(
-          metadataData._id,
-          new MetadataEntity({
-            externalId: falResult.url,
-            prompt: promptData._id,
-          }),
-        );
-
-        for (let i = 1; i < outputs; i++) {
-          const {
-            metadataData: additionalMetadata,
-            ingredientData: additionalIngredient,
-          } = await this.sharedService.saveDocuments(user, {
-            ...createImageDto,
-            brand: brand._id,
-            category: IngredientCategory.IMAGE,
-            extension: MetadataExtension.JPG,
-            model,
-            organization: publicMetadata.organization,
-            parent: ingredientData.parent,
-            prompt: promptData._id,
-            status: IngredientStatus.PROCESSING,
-          });
-
-          const additionalResult = await this.falService.generateImage(
-            model,
-            buildFalInput(),
-          );
-
-          await Promise.all([
-            this.metadataService.patch(
-              additionalMetadata._id,
-              new MetadataEntity({
-                externalId: additionalResult.url,
-                prompt: promptData._id,
-              }),
-            ),
-            this.imagesService.patch(additionalIngredient._id, {
-              prompt: promptData._id,
-            }),
-          ]);
-
-          const additionalActivity = await this.activitiesService.create(
-            new ActivityEntity({
-              brand: brand._id,
-              entityId: additionalIngredient._id,
-              entityModel: ActivityEntityModel.INGREDIENT,
-              key: ActivityKey.IMAGE_PROCESSING,
-              organization: publicMetadata.organization,
-              source: ActivitySource.IMAGE_GENERATION,
-              user: publicMetadata.user,
-              value: JSON.stringify({
-                ingredientId: additionalIngredient._id.toString(),
-                model,
-                type: 'generation',
-              }),
-            }),
-          );
-
-          await this.websocketService.publishBackgroundTaskUpdate({
-            activityId: additionalActivity._id.toString(),
-            label: 'Image Generation',
-            progress: 0,
-            room: getUserRoomName(user.id),
-            status: 'processing',
-            taskId: additionalIngredient._id.toString(),
-            userId: user.id,
-          });
-
-          pendingIngredientIds.push(additionalIngredient._id.toString());
-        }
-
-        return falResult.url;
-      })().catch(async (error: unknown) => {
-        this.loggerService.error(`${url} failed`, error);
-        const errorMessage = getErrorMessage(error);
-
-        await this.failedGenerationService.handleFailedImageGeneration(
-          this.imagesService,
-          ingredientData._id,
-          websocketUrl,
-          publicMetadata,
-          getUserRoomName(user.id),
-          errorMessage,
-        );
-        throw error;
-      });
-
-      generationPromise.catch(() => {
-        // Error already handled in the catch block above
-      });
-    }
-
-    // Leonardo
-    if (isLeonardoAI) {
-      url = 'LeonardoAIService generateImage';
-
-      // Create async generation promise - don't await unless waitForCompletion
-      const generationPromise = this.leonardoaiService
-        .generateImage(promptData.original, {
-          ...createImageDto,
-          height,
-          style: style || 'realistic',
-          width,
-        })
-        .then(async (generationId) => {
-          if (!generationId) {
-            throw new Error('No generation ID returned from LeonardoAI');
-          }
-
-          await this.metadataService.patch(
-            metadataData._id,
-            new MetadataEntity({
-              externalId: generationId,
-            }),
-          );
-
-          return generationId;
-        })
-        .catch(async (error: unknown) => {
-          this.loggerService.error(`${url} failed`, error);
-          const errorMessage = getErrorMessage(error);
-
-          await this.failedGenerationService.handleFailedImageGeneration(
-            this.imagesService,
-            ingredientData._id,
-            websocketUrl,
-            publicMetadata,
-            getUserRoomName(user.id),
-            errorMessage,
-          );
-          throw error;
-        });
-
-      if (waitForCompletion) {
-        // Wait for the generation to complete
-        try {
-          await generationPromise;
-
-          const completedIngredient =
-            await this.pollingService.waitForIngredientCompletion(
-              ingredientData._id.toString(),
-              180000, // 3 minutes timeout
-              2000, // 2 seconds poll interval
-              [
-                PopulatePatterns.promptFull,
-                PopulatePatterns.metadataFull,
-                PopulatePatterns.brandMinimal,
-              ],
-            );
-
-          return serializeSingle(
-            request,
-            IngredientSerializer,
-            completedIngredient,
-          );
-        } catch (error: unknown) {
-          if (error instanceof Error && error.name === 'PollingTimeoutError') {
-            const ingredient = await this.imagesService.findOne(
-              { _id: ingredientData._id },
-              [
-                PopulatePatterns.promptFull,
-                PopulatePatterns.metadataFull,
-                PopulatePatterns.brandMinimal,
-              ],
-            );
-
-            if (ingredient) {
-              throw new HttpException(
-                {
-                  detail: `Image generation did not complete within 3 minutes. Current status: ${ingredient.status}`,
-                  title: 'Generation timeout',
-                },
-                HttpStatus.GATEWAY_TIMEOUT,
-              );
-            }
-          }
-          throw error;
-        }
-      } else {
-        // Return immediately - generation runs in background
-        // Attach empty catch to prevent unhandled promise rejection
-        generationPromise.catch(() => {
-          // Error already handled in the catch block above
-        });
-      }
-    }
-
-    if (isReplicate) {
-      url = 'ReplicateService generateImage';
-
-      // Build provider-specific prompt using universal prompt builder with template support
-      // NOTE: This must complete before we can start generation
-      const { input: promptParams } =
-        await this.promptBuilderService.buildPrompt(
-          model,
-          {
-            blacklist: createImageDto.blacklist,
-            brand: promptBuilderBrand,
-            branding: brandPromptBranding,
-            brandingMode: createImageDto.brandingMode,
-            camera: createImageDto.camera,
-            fontFamily: createImageDto.fontFamily,
-            height,
-            isBrandingEnabled: createImageDto.isBrandingEnabled,
-            lens: createImageDto.lens,
-            lighting: createImageDto.lighting,
-            // Use model's category from DB (set by ModelsGuard), fallback to IMAGE
-            modelCategory: ModelCategory.IMAGE,
-            mood: createImageDto.mood,
-            outputs,
-            prompt: promptData.original,
-            // Template support
-            promptTemplate: createImageDto.promptTemplate,
-            references: referenceImageUrls,
-            scene: createImageDto.scene,
-            seed: createImageDto.seed,
-            style: style || createImageDto.style || 'realistic',
-            tags: createImageDto.tags?.map((tag) => tag.toString()) || [],
-            useTemplate: createImageDto.useTemplate,
-            width,
-          },
-          publicMetadata.organization,
-        );
-
-      const destination = model;
-      const modelCapability = MODEL_OUTPUT_CAPABILITIES[destination];
-      const isBatchSupported = modelCapability?.isBatchSupported ?? false;
-
-      // Track all placeholder ingredient IDs - start with first one
-      const replicatePendingIngredientIds: string[] = [
-        ingredientData._id.toString(),
-      ];
-
-      // Create async generation promise - don't await unless waitForCompletion
-      const generationPromise = this.replicateService
-        .generateTextToImage(destination, promptParams)
-        .then(async (generationId) => {
-          if (!generationId) {
-            throw new Error('No generation ID returned from Replicate');
-          }
-
-          // Handle multiple outputs differently for batch-capable models vs others
-          if (isBatchSupported && outputs > 1) {
-            // Batch-capable models: single API call with num_outputs
-            await this.metadataService.patch(
-              metadataData._id,
-              new MetadataEntity({
-                externalId: `${generationId}_0`,
-              }),
-            );
-
-            // Create additional placeholder documents for remaining outputs
-            const additionalDocuments = await Promise.all(
-              Array.from({ length: outputs - 1 }, () => {
-                return this.sharedService.saveDocuments(user, {
-                  ...createImageDto,
-                  brand: brand._id,
-                  category: IngredientCategory.IMAGE,
-                  extension: MetadataExtension.JPG,
-                  model,
-                  organization: publicMetadata.organization,
-                  parent: ingredientData.parent,
-                  prompt: promptData._id,
-                  status: IngredientStatus.PROCESSING,
-                });
-              }),
-            );
-
-            // Update metadata and images in parallel
-            await Promise.all(
-              additionalDocuments.flatMap(
-                (
-                  { metadataData: addMeta, ingredientData: addIngredient },
-                  index,
-                ) => {
-                  const i = index + 1;
-                  return [
-                    this.metadataService.patch(
-                      addMeta._id,
-                      new MetadataEntity({
-                        externalId: `${generationId}_${i}`,
-                      }),
-                    ),
-                    this.imagesService.patch(addIngredient._id, {
-                      prompt: promptData._id,
-                    }),
-                  ];
-                },
-              ),
-            );
-
-            // Create activities for each additional placeholder (batch model path)
-            await Promise.all(
-              additionalDocuments.map(({ ingredientData: addIngredient }) =>
-                this.activitiesService
-                  .create(
-                    new ActivityEntity({
-                      brand: brand._id,
-                      entityId: addIngredient._id,
-                      entityModel: ActivityEntityModel.INGREDIENT,
-                      key: ActivityKey.IMAGE_PROCESSING,
-                      organization: publicMetadata.organization,
-                      source: ActivitySource.IMAGE_GENERATION,
-                      user: publicMetadata.user,
-                      value: JSON.stringify({
-                        ingredientId: addIngredient._id.toString(),
-                        model,
-                        type: 'generation',
-                      }),
-                    }),
-                  )
-                  .then((newActivity) =>
-                    this.websocketService.publishBackgroundTaskUpdate({
-                      activityId: newActivity._id.toString(),
-                      label: 'Image Generation',
-                      progress: 0,
-                      room: getUserRoomName(user.id),
-                      status: 'processing',
-                      taskId: addIngredient._id.toString(),
-                      userId: user.id,
-                    }),
-                  ),
-              ),
-            );
-
-            // Push additional ingredient IDs to tracking array for polling
-            additionalDocuments.forEach(({ ingredientData: addIngredient }) => {
-              replicatePendingIngredientIds.push(addIngredient._id.toString());
-            });
-
-            this.loggerService.log(
-              'Created multiple placeholders for batch-capable model multi-output',
-              {
-                generationId,
-                isBatchSupported: true,
-                model: destination,
-                outputs,
-              },
-            );
-          } else if (outputs > 1) {
-            // Non-batch models: make multiple separate API calls
-            await this.metadataService.patch(
-              metadataData._id,
-              new MetadataEntity({
-                externalId: generationId,
-              }),
-            );
-
-            // Make additional API calls for remaining outputs
-            // NOTE: Sequential because each needs unique generationId
-            for (let i = 1; i < outputs; i++) {
-              const {
-                metadataData: additionalMetadata,
-                ingredientData: additionalIngredient,
-              } = await this.sharedService.saveDocuments(user, {
-                ...createImageDto,
-                brand: brand._id,
-                category: IngredientCategory.IMAGE,
-                extension: MetadataExtension.JPG,
-                model,
-                organization: publicMetadata.organization,
-                parent: ingredientData.parent,
-                prompt: promptData._id,
-                status: IngredientStatus.PROCESSING,
-              });
-
-              // Make separate API call for each output
-              const additionalGenerationId =
-                await this.replicateService.generateTextToImage(
-                  destination,
-                  promptParams,
-                );
-
-              await Promise.all([
-                this.metadataService.patch(
-                  additionalMetadata._id,
-                  new MetadataEntity({
-                    externalId: additionalGenerationId,
-                  }),
-                ),
-                this.imagesService.patch(additionalIngredient._id, {
-                  prompt: promptData._id,
-                }),
-              ]);
-
-              // Create activity for this additional output (non-batch path)
-              const additionalActivity = await this.activitiesService.create(
-                new ActivityEntity({
-                  brand: brand._id,
-                  entityId: additionalIngredient._id,
-                  entityModel: ActivityEntityModel.INGREDIENT,
-                  key: ActivityKey.IMAGE_PROCESSING,
-                  organization: publicMetadata.organization,
-                  source: ActivitySource.IMAGE_GENERATION,
-                  user: publicMetadata.user,
-                  value: JSON.stringify({
-                    ingredientId: additionalIngredient._id.toString(),
-                    model,
-                    type: 'generation',
-                  }),
-                }),
-              );
-
-              await this.websocketService.publishBackgroundTaskUpdate({
-                activityId: additionalActivity._id.toString(),
-                label: 'Image Generation',
-                progress: 0,
-                room: getUserRoomName(user.id),
-                status: 'processing',
-                taskId: additionalIngredient._id.toString(),
-                userId: user.id,
-              });
-
-              // Push this additional ingredient ID to tracking array for polling
-              replicatePendingIngredientIds.push(
-                additionalIngredient._id.toString(),
-              );
-            }
-
-            this.loggerService.log(
-              'Created multiple API calls for non-batch model multi-output',
-              {
-                isBatchSupported: false,
-                model: destination,
-                outputs,
-              },
-            );
-          } else {
-            // Single output - use original external ID
-            await this.metadataService.patch(
-              metadataData._id,
-              new MetadataEntity({
-                externalId: generationId,
-              }),
-            );
-          }
-
-          return generationId;
-        })
-        .catch(async (error: unknown) => {
-          this.loggerService.error(`${url} failed`, error);
-          const errorMessage = getErrorMessage(error);
-
-          await this.failedGenerationService.handleFailedImageGeneration(
-            this.imagesService,
-            ingredientData._id,
-            websocketUrl,
-            publicMetadata,
-            getUserRoomName(user.id),
-            errorMessage,
-          );
-          throw error;
-        });
-
-      if (waitForCompletion) {
-        // Wait for the generation to complete
-        try {
-          await generationPromise;
-
-          const completedIngredients =
-            await this.pollingService.waitForMultipleIngredientsCompletion(
-              replicatePendingIngredientIds,
-              180_000, // 3 minutes timeout
-              2_000, // 2 seconds poll interval
-              [
-                PopulatePatterns.promptFull,
-                PopulatePatterns.metadataFull,
-                PopulatePatterns.brandMinimal,
-              ],
-            );
-
-          return serializeSingle(
-            request,
-            IngredientSerializer,
-            completedIngredients[0],
-          );
-        } catch (error: unknown) {
-          if (error instanceof Error && error.name === 'PollingTimeoutError') {
-            const ingredient = await this.imagesService.findOne(
-              { _id: ingredientData._id },
-              [
-                PopulatePatterns.promptFull,
-                PopulatePatterns.metadataFull,
-                PopulatePatterns.brandMinimal,
-              ],
-            );
-
-            if (ingredient) {
-              throw new HttpException(
-                {
-                  detail: `Image generation did not complete within 3 minutes. Current status: ${ingredient.status}`,
-                  title: 'Generation timeout',
-                },
-                HttpStatus.GATEWAY_TIMEOUT,
-              );
-            }
-          }
-          throw error;
-        }
-      } else {
-        // Return immediately - generation runs in background
-        // Attach empty catch to prevent unhandled promise rejection
-        generationPromise.catch(() => {
-          // Error already handled in the catch block above
-        });
-      }
-    }
-
-    // For non-Replicate providers, handle waitForCompletion if needed
-    if (
-      waitForCompletion &&
-      !isKlingAI &&
-      !isLeonardoAI &&
-      !isFal &&
-      !isReplicate
-    ) {
-      // This shouldn't happen, but handle it gracefully
-      this.loggerService.warn(
-        'waitForCompletion requested for unsupported provider',
-        {
-          ingredientId: ingredientData._id,
-          model,
-        },
-      );
-    }
-
-    // For non-Replicate providers, return single ID
-    return serializeSingle(request, IngredientSerializer, {
-      ...ingredientData,
-      pendingIngredientIds,
-    });
   }
 
   /**
@@ -1445,41 +302,6 @@ export class ImagesOperationsController {
         frames: frameResults,
       },
     };
-  }
-
-  /**
-   * Calculate dynamic image cost based on model pricing type.
-   * Mirrors the CreditsGuard.calculateDynamicCost logic for image models.
-   */
-  private calculateDynamicImageCost(
-    model: {
-      cost?: number;
-      pricingType?: PricingType;
-      costPerUnit?: number;
-      minCost?: number;
-    },
-    width: number,
-    height: number,
-  ): number {
-    const pricingType = model.pricingType || PricingType.FLAT;
-    let baseCost = model.cost || 0;
-
-    if (
-      pricingType === PricingType.PER_MEGAPIXEL &&
-      width &&
-      height &&
-      model.costPerUnit
-    ) {
-      const megapixels = (width * height) / 1_000_000;
-      baseCost = Math.ceil(megapixels * model.costPerUnit);
-    }
-
-    const minCost = model.minCost || 0;
-    if (minCost > 0 && baseCost < minCost) {
-      baseCost = minCost;
-    }
-
-    return baseCost;
   }
 
   private getRefId(
